pyproject.toml
[project]
name = "slow-ai"
version = "0.1.0"
description = "Terminal-based interview agent that produces a validated problem brief"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.6",
    "pydantic>=2.0",
    "pydantic-ai[google]>=1.70.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "streamlit>=1.40.0",
    "nest-asyncio>=1.6.0",
    "gitpython>=3.1",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "streamlit-flow-component>=1.6.1",
    "pdfplumber>=0.11",
    "pandas>=2.0",
    "bandit>=1.7",
    "pyyaml>=6.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1",
    "python-docx>=1.0",
    "mistune>=3.0",
]

[project.scripts]
slow-ai = "slow_ai.cli:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "types-PyYAML>=6.0",
    "types-beautifulsoup4>=4.12",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
no_strict_optional = true
# pydantic-ai stubs don't fully type Agent(output_type=...) — suppress those modules
[[tool.mypy.overrides]]
module = [
    "slow_ai.agents.*",
    "slow_ai.skills.resolver",
    "slow_ai.skills.synthesizer",
    "slow_ai.tools.code_generation",
    "slow_ai.config",
    "slow_ai.research.runner",
]
ignore_errors = true

[[tool.mypy.overrides]]
# Starlette/Jinja2 stubs don't match the TemplateResponse(name, context) call style
module = ["app.api.*"]
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slow_ai"]
